+++ contracts/dataFlow/FixedPriceFlow.sol
@@ -6,20 +6,16 @@ import "../market/FixedPrice.sol";
 contract FixedPriceFlow is Flow {
     FixedPrice public immutable market;
 
-    error NotEnoughFee(uint256 price, uint256 amount, uint256 paid);
+    error NotEnoughFee(uint price, uint amount, uint paid);
 
-    constructor(
-        address book_,
-        uint256 blocksPerEpoch_,
-        uint256 deployDelay_
-    ) Flow(book_, blocksPerEpoch_, deployDelay_) {
+    constructor(address book_, uint blocksPerEpoch_, uint deployDelay_) Flow(book_, blocksPerEpoch_, deployDelay_) {
         market = FixedPrice(payable(address(book.market())));
     }
 
-    function _beforeSubmit(uint256 sectors) internal override {
-        uint256 price = market.pricePerSector();
-        uint256 fee = sectors * price;
-        uint256 paid = address(this).balance;
+    function _beforeSubmit(uint sectors) internal override {
+        uint price = market.pricePerSector();
+        uint fee = sectors * price;
+        uint paid = address(this).balance;
 
         if (fee > address(this).balance) {
             revert NotEnoughFee(price, sectors, paid);
