+++ contracts/interfaces/IMarket.sol
@@ -2,9 +2,5 @@
 pragma solidity >=0.8.0 <0.9.0;
 
 interface IMarket {
-    function chargeFee(
-        uint256 beforeLength,
-        uint256 uploadSectors,
-        uint256 paddingSectors
-    ) external;
+    function chargeFee(uint beforeLength, uint uploadSectors, uint paddingSectors) external;
 }
