+++ contracts/miner/RecallRange.sol
@@ -5,8 +5,8 @@ import "@openzeppelin/contracts/utils/math/Math.sol";
 import "../utils/ZgsSpec.sol";
 
 struct RecallRange {
-    uint256 startPosition;
-    uint256 mineLength;
+    uint startPosition;
+    uint mineLength;
     uint64 shardId;
     uint64 shardMask;
 }
@@ -14,38 +14,22 @@ struct RecallRange {
 library RecallRangeLib {
     using RecallRangeLib for RecallRange;
 
-    function check(RecallRange memory range, uint256 maxEndPosition)
-        internal
-        pure
-    {
-        require(
-            range.startPosition % SECTORS_PER_PRICE == 0,
-            "Start position is not aligned"
-        );
+    function check(RecallRange memory range, uint maxEndPosition) internal pure {
+        require(range.startPosition % SECTORS_PER_PRICE == 0, "Start position is not aligned");
 
-        require(
-            range.startPosition + range.mineLength <= maxEndPosition,
-            "Mining range overflow"
-        );
+        require(range.startPosition + range.mineLength <= maxEndPosition, "Mining range overflow");
 
-        uint256 maxMineLength = MAX_MINING_LENGTH * range.numShards();
+        uint maxMineLength = MAX_MINING_LENGTH * range.numShards();
         require(range.mineLength <= maxMineLength, "Mining range too long");
 
-        uint256 requiredLength = Math.min(maxEndPosition, maxMineLength);
+        uint requiredLength = Math.min(maxEndPosition, maxMineLength);
         require(range.mineLength >= requiredLength, "Mining range too short");
 
-        require(
-            range.shardId & range.shardMask == 0,
-            "Masked bits should be zero"
-        );
+        require(range.shardId & range.shardMask == 0, "Masked bits should be zero");
     }
 
     // Compute bit `1` in shardMask, while assuming there is only a few zeros.
-    function numShards(RecallRange memory range)
-        internal
-        pure
-        returns (uint256)
-    {
+    function numShards(RecallRange memory range) internal pure returns (uint) {
         uint64 negMask = ~range.shardMask;
         uint8 answer = 0;
         while (negMask > 0) {
@@ -56,31 +40,14 @@ library RecallRangeLib {
     }
 
     function digest(RecallRange memory range) internal pure returns (bytes32) {
-        return
-            keccak256(
-                abi.encode(
-                    range.startPosition,
-                    range.mineLength,
-                    range.shardId,
-                    range.shardMask
-                )
-            );
+        return keccak256(abi.encode(range.startPosition, range.mineLength, range.shardId, range.shardMask));
     }
 
-    function recallChunk(RecallRange memory range, bytes32 seed)
-        internal
-        pure
-        returns (uint256)
-    {
-        uint256 originChunkOffset = uint256(seed) %
-            (range.mineLength / SECTORS_PER_LOAD);
-        uint64 chunkOffset = (uint64(originChunkOffset) & range.shardMask) |
-            range.shardId;
+    function recallChunk(RecallRange memory range, bytes32 seed) internal pure returns (uint) {
+        uint originChunkOffset = uint(seed) % (range.mineLength / SECTORS_PER_LOAD);
+        uint64 chunkOffset = (uint64(originChunkOffset) & range.shardMask) | range.shardId;
 
-        require(
-            chunkOffset * SECTORS_PER_LOAD <= range.mineLength,
-            "Recall position out of bound"
-        );
+        require(chunkOffset * SECTORS_PER_LOAD <= range.mineLength, "Recall position out of bound");
 
         return range.startPosition + chunkOffset * SECTORS_PER_LOAD;
     }
